Drink/Food/Product models, Add orders and see basket state

+++ BlazorAthenaFrontend/Models/Drink.cs
@@ -3,6 +3,6 @@
     public class Drink
     {
         public int ID { get; set; }
-        public int? AlcoholPercentage { get; set; }
+        public decimal? AlcoholPercentage { get; set; }
     }
 }
